feat: setup mobile style (#160)
* feat: setup mobile style

* chore: lint

* chore: fix loading error

* refactor: fix style issues

* refactor: change project layout

* refactor: project layout

* refactor: move dataStorage into state

* refactor: add title to mobile home page

* refactor: fix bug in invoice input and fix styles

* refactor: fix styles now they are less shitty :)

* refactor: fix some bugs

* Update src/components/navigationbar/desktopnavigationbar.js


* chore: fix bugs and lint code

* refactor: fix issue where id dosent exsist on swapResponse

* refactor: fix reverse swap flow

* chore: lint
