+++ src/actions/index.js
@@ -2,5 +2,5 @@ import { generateKeys } from './keys';
 import history from '../constants/history';
 import Navigation from './navigation';
 
-export const navigation = new Navigation(history);
-export { generateKeys };
+const navigation = new Navigation(history);
+export { generateKeys, navigation };
