+++ src/constants/actions/index.js
@@ -37,3 +37,10 @@ export const SET_REFUND_DESTINATION = 'SET_REFUND_DESTINATION';
 export const SET_REFUND_TRANSACTION_HASH = 'SET_REFUND_TRANSACTION_HASH';
 export const REFUND_REQUEST = 'REFUND_REQUEST';
 export const REFUND_RESPONSE = 'REFUND_RESPONSE';
+
+/**
+ * Data Storage
+ */
+export const SET_ID = 'SET_ID';
+export const SET_ASSET = 'SET_ASSET';
+export const CLEAR_DATA_STORAGE = 'CLEAR_DATA_STORAGE';
