+++ src/views/landingpage/index.js
@@ -1,8 +1,19 @@
+import React, { lazy } from 'react';
 import { connect } from 'react-redux';
-import LandingPage from './landingpage';
-import { initSwap } from '../swap/swapActions';
-import * as actions from './landingpageActions';
-import { initReverseSwap } from '../reverse/reverseActions';
+import PlatformSelector from '../../hoc/platformSelector';
+import { initSwap } from '../../actions/swapActions';
+import * as actions from '../../actions/landingPageActions';
+import { initReverseSwap } from '../../actions/reverseActions';
+
+const LandingPageDesktop = lazy(() => import('./landingPageDesktop'));
+const LandingPageMobile = lazy(() => import('./landingPageMobile'));
+
+const LandingPage = props => (
+  <PlatformSelector
+    mobile={<LandingPageMobile {...props} />}
+    desktop={<LandingPageDesktop {...props} />}
+  />
+);
 
 const mapStateToProps = state => ({
   warnings: state.landingpageReducer.warnings,
