--- src/views/reversetimelock/dataStorage.js
@@ -1,9 +1,0 @@
-class DataStorage {
-  static swapInfo = {
-    id: '',
-    asset: '',
-    amount: 0,
-  };
-}
-
-export default DataStorage;
