+++ src/views/reversetimelock/index.js
@@ -1,9 +1,18 @@
 import { connect } from 'react-redux';
-import ReverseSwapTimelockExpired from './reverseSwapTimelockExpired';
+import { dataStorageClear } from '../../actions/datastorageActions';
+import ReverseSwapTimelockExpired from './ReverseSwapTimelockExpired';
 
-const mapDispatchToProps = () => ({});
+const mapStateToProps = state => ({
+  id: state.dataStorage.id,
+  asset: state.dataStorage.asset,
+  amount: state.dataStorage.amount,
+});
+
+const mapDispatchToProps = dispatch => ({
+  dataStorageClear: () => dispatch(dataStorageClear()),
+});
 
 export default connect(
-  null,
+  mapStateToProps,
   mapDispatchToProps
 )(ReverseSwapTimelockExpired);
