+++ src/views/swap/swap.js
@@ -9,7 +9,7 @@ import Confetti from '../../components/confetti';
 import BackGround from '../../components/background';
 import StepsWizard from '../../components/stepswizard';
 import { InputInvoice, SendTransaction, DownloadRefund } from './steps';
-import { navigation } from '../../action';
+import { navigation } from '../../actions';
 
 const styles = () => ({
   wrapper: {
@@ -119,7 +119,9 @@ class Swap extends Component {
                     text={`Next`}
                     loadingText={'Invalid invoice'}
                     onPress={() => {
-                      startSwap(swapInfo, props.nextStage);
+                      if (swapInfo.invoice) {
+                        startSwap(swapInfo, props.nextStage);
+                      }
                     }}
                   />
                 )}
